Swap airline and base16-vim for nvim-lualine and nvim-base16
Having not spent much time configuring airline, didn't feel I had much
to lose jumping on a neovim / lua take on status lines. This needed the
RRethy/nvim-base16 plugin for the base16 colorschemes; so using that
instead of the more familiar "supported" base16-vim plugin.

Think this might only cause an issue for 16 colour terminals, rather
than modern, true color (?) terminals, which is unlikely to be an issue
for my development in neovim