Remove outdated null_ls plugins
The none-ls project has pruned some older plugins that didn't get a lot
of support. One of them; the whitespace trimming, can be done with
editorconfig which neovim now supports, so this adds that back in

The other two, eslint_d and shellcheck, have alternatives, so I can try
and install those when I'm next working in those file types.

Ref: https://github.com/nvimtools/none-ls.nvim/issues/58